requirements.txt
fastapi==0.70.1
numpy==1.21.4
pandas==1.3.4
pydantic==1.8.2
pytz==2021.3
selenium==4.1.0
SQLAlchemy==1.4.28
tda_api==1.5